requirements.txt
Django==4.2.20
djangorestframework==3.14.0
djangorestframework-simplejwt
django-cors-headers==4.3.1
mysqlclient==2.2.4
gunicorn==21.2.0
python-dotenv==1.0.1
django-filter==23.5
django-cleanup==8.1.0
Pillow==10.2.0
openpyxl==3.1.2
django-storages
boto3
